pyproject.toml
[tool.poetry]
name = "cvna_gsc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.23.0"
certifi = "^2020.4.5.1"
google-api-python-client = "^1.8.3"
google-auth = "^1.15.0"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
httplib2 = "^0.18.1"
idna = "^2.6"
oauth2client = "^4.1.2"
oauthlib = "^3.1.0"
pyasn1 = "^0.4.2"
pyasn1-modules = "^0.2.1"
requests-oauthlib = "^0.8.0"
rsa = "^3.4.2"
uritemplate = "^3.0.0"
urllib3 = "^1.25.9"
python-dateutil = "^2.8.1"
pandas = "^1.2.3"

[tool.poetry.dev-dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
